@@ -9,7 +9,8 @@
 import TWSModels
 
 /// Events that are sent to ``TWSManager`` regarding updates
-public enum TWSStreamEvent {
+public enum TWSStreamEvent: Sendable {
+
     /// This event is sent when a project from the universal link is processed
     /// - Parameter TWSSharedSnippetBundle: A snippet that should be preselected (opened) along with organization and project info
     case universalLinkSnippetLoaded(TWSSharedSnippetBundle)
