@@ -32,9 +32,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug", xcconfig: .relativeToRoot("config/TWSDemo_dev.xcconfig")),
-                    .release(name: "Staging", xcconfig: .relativeToRoot("config/TWSDemo_staging.xcconfig")),
-                    .release(name: "Release", xcconfig: .relativeToRoot("config/TWSDemo_release.xcconfig"))
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDemo_dev.xcconfig")),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDemo_staging.xcconfig")),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDemo_release.xcconfig"))
                 ],
                 defaultSettings: .recommended(excluding: [
                     "CODE_SIGN_IDENTITY",
@@ -57,6 +57,17 @@ let project = Project(
                 configurations: [
                     .debug(
                         name: "Debug",
+                        settings: ["SWIFT_VERSION": "6.0"],
+                        xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
+                    ),
+                    .release(
+                        name: "Staging",
+                        settings: ["SWIFT_VERSION": "6.0"],
+                        xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
+                    ),
+                    .release(
+                        name: "Release",
+                        settings: ["SWIFT_VERSION": "6.0"],
                         xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
                     )
                 ]
@@ -77,9 +88,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug", xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
-                    .release(name: "Staging", xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
-                    .release(name: "Release", xcconfig: .relativeToRoot("config/TWSDist.xcconfig"))
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig"))
                 ]
             )
         ),
@@ -95,9 +106,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig"))
                 ]
             )
         ),
@@ -117,9 +128,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"])
                 ]
             )
         ),
@@ -138,9 +149,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"])
                 ]
             )
         ),
@@ -159,6 +170,17 @@ let project = Project(
                 configurations: [
                     .debug(
                         name: "Debug",
+                        settings: ["SWIFT_VERSION": "6.0"],
+                        xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
+                    ),
+                    .release(
+                        name: "Staging",
+                        settings: ["SWIFT_VERSION": "6.0"],
+                        xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
+                    ),
+                    .release(
+                        name: "Release",
+                        settings: ["SWIFT_VERSION": "6.0"],
                         xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
                     )
                 ]
@@ -178,9 +200,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"])
                 ]
             )
         ),
@@ -198,9 +220,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"])
                 ]
             )
         ),
@@ -215,9 +237,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug", xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
-                    .release(name: "Staging", xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
-                    .release(name: "Release", xcconfig: .relativeToRoot("config/TWSDist.xcconfig"))
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig")),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"], xcconfig: .relativeToRoot("config/TWSDist.xcconfig"))
                 ]
             )
         ),
@@ -234,9 +256,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"])
                 ]
             )
         ),
@@ -255,9 +277,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"])
                 ]
             )
         ),
@@ -274,9 +296,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"])
                 ]
             )
         ),
@@ -295,6 +317,17 @@ let project = Project(
                 configurations: [
                     .debug(
                         name: "Debug",
+                        settings: ["SWIFT_VERSION": "6.0"],
+                        xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
+                    ),
+                    .release(
+                        name: "Staging",
+                        settings: ["SWIFT_VERSION": "6.0"],
+                        xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
+                    ),
+                    .release(
+                        name: "Release",
+                        settings: ["SWIFT_VERSION": "6.0"],
                         xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
                     )
                 ]
@@ -314,9 +347,9 @@ let project = Project(
             ],
             settings: .settings(
                 configurations: [
-                    .debug(name: "Debug"),
-                    .release(name: "Staging"),
-                    .release(name: "Release")
+                    .debug(name: "Debug", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Staging", settings: ["SWIFT_VERSION": "6.0"]),
+                    .release(name: "Release", settings: ["SWIFT_VERSION": "6.0"])
                 ]
             )
         ),
@@ -335,6 +368,17 @@ let project = Project(
                 configurations: [
                     .debug(
                         name: "Debug",
+                        settings: ["SWIFT_VERSION": "6.0"],
+                        xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
+                    ),
+                    .release(
+                        name: "Staging",
+                        settings: ["SWIFT_VERSION": "6.0"],
+                        xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
+                    ),
+                    .release(
+                        name: "Release",
+                        settings: ["SWIFT_VERSION": "6.0"],
                         xcconfig: .relativeToRoot("config/TWSDemo_tests.xcconfig")
                     )
                 ]
