@@ -119,21 +119,21 @@ struct SettingsView: View {
             viewModel.setSource(manager: twsViewModel.manager, source: newValue)
         }
     }
-}
 
-private func shareLogsReport(_ reportUrl: URL?) {
-    if let reportUrl {
-        let activityVC = UIActivityViewController(activityItems: [reportUrl], applicationActivities: nil)
-        activityVC.completionWithItemsHandler = { (_, _, _, _) in
-            do {
-                try FileManager.default.removeItem(atPath: reportUrl.path())
-            } catch {
-                print("Unable to delete the logs file after sharing")
+    private func shareLogsReport(_ reportUrl: URL?) {
+        if let reportUrl {
+            let activityVC = UIActivityViewController(activityItems: [reportUrl], applicationActivities: nil)
+            activityVC.completionWithItemsHandler = { (_, _, _, _) in
+                do {
+                    try FileManager.default.removeItem(atPath: reportUrl.path())
+                } catch {
+                    print("Unable to delete the logs file after sharing")
+                }
             }
-        }
-        if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
-           let rootVC = windowScene.windows.first?.rootViewController {
-            rootVC.present(activityVC, animated: true, completion: nil)
+            if let windowScene = UIApplication.shared.connectedScenes.first as? UIWindowScene,
+               let rootVC = windowScene.windows.first?.rootViewController {
+                rootVC.present(activityVC, animated: true, completion: nil)
+            }
         }
     }
 }
