@@ -47,7 +47,7 @@ final class SnippetsTests: XCTestCase {
             initialState: state,
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: snippets)}
+                $0.api.getProject = { [socketURL] _ in TWSProject(listenOn: socketURL, snippets: snippets)}
             }
         )
 
@@ -86,7 +86,7 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: snippets)}
+                $0.api.getProject = { [socketURL] _ in TWSProject(listenOn: socketURL, snippets: snippets)}
             }
         )
 
@@ -100,8 +100,8 @@ final class SnippetsTests: XCTestCase {
 
         // Send for the second time without one element. Snippet should be removed from state
 
-        store.dependencies.api.getProject = { _ in
-            TWSProject(listenOn: self.socketURL, snippets: [snippets[1], snippets[2]])
+        store.dependencies.api.getProject = { [socketURL] _ in
+            TWSProject(listenOn: socketURL, snippets: [snippets[1], snippets[2]])
         }
 
         await store.send(.business(.load)).finish()
@@ -126,8 +126,8 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in
-                    TWSProject(listenOn: self.socketURL, snippets: [snippets[0], snippets[2]])
+                $0.api.getProject = { [socketURL] _ in
+                    TWSProject(listenOn: socketURL, snippets: [snippets[0], snippets[2]])
                 }
             }
         )
@@ -141,7 +141,7 @@ final class SnippetsTests: XCTestCase {
         }
 
         // Send for the second time with new element. Snippet should be added in right order
-        store.dependencies.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: snippets)}
+        store.dependencies.api.getProject = { [socketURL] _ in TWSProject(listenOn: socketURL, snippets: snippets)}
 
         await store.send(.business(.load)).finish()
         await store.receive(\.business.projectLoaded.success) {
@@ -167,7 +167,7 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: snippets)}
+                $0.api.getProject = { [socketURL] _ in TWSProject(listenOn: socketURL, snippets: snippets)}
             }
         )
 
@@ -181,8 +181,8 @@ final class SnippetsTests: XCTestCase {
 
         // Send response for the second time but change the order
 
-        store.dependencies.api.getProject = { _ in
-            TWSProject(listenOn: self.socketURL, snippets: [snippets[1], snippets[2], snippets[0]])
+        store.dependencies.api.getProject = { [socketURL] _ in
+            TWSProject(listenOn: socketURL, snippets: [snippets[1], snippets[2], snippets[0]])
         }
 
         await store.send(.business(.load)).finish()
@@ -211,8 +211,8 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in
-                    TWSProject(listenOn: self.socketURL, snippets: [snippets[0], snippets[1], snippets[2]])
+                $0.api.getProject = { [socketURL] _ in
+                    TWSProject(listenOn: socketURL, snippets: [snippets[0], snippets[1], snippets[2]])
                 }
             }
         )
@@ -227,8 +227,8 @@ final class SnippetsTests: XCTestCase {
 
         // Send response for the second time but remove some and add some
 
-        store.dependencies.api.getProject = { _ in
-            TWSProject(listenOn: self.socketURL, snippets: [snippets[0], snippets[2], snippets[3]])
+        store.dependencies.api.getProject = { [socketURL] _ in
+            TWSProject(listenOn: socketURL, snippets: [snippets[0], snippets[2], snippets[3]])
         }
 
         await store.send(.business(.load)).finish()
@@ -244,7 +244,7 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsObserverFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in .init(listenOn: self.socketURL, snippets: [])}
+                $0.api.getProject = { [socketURL] _ in .init(listenOn: socketURL, snippets: [])}
                 $0.socket.get = { _, _ in .init() }
                 $0.socket.connect = { _ in .makeStream().stream }
                 $0.socket.closeConnection = { _ in }
@@ -275,7 +275,7 @@ final class SnippetsTests: XCTestCase {
             initialState: TWSSnippetsFeature.State(configuration: configuration),
             reducer: { TWSSnippetsFeature() },
             withDependencies: {
-                $0.api.getProject = { _ in TWSProject(listenOn: self.socketURL, snippets: [])}
+                $0.api.getProject = { [socketURL] _ in TWSProject(listenOn: socketURL, snippets: [])}
             }
         )
 
